feat: Add SpotDTO and update Event and Ticket models
This commit adds the SpotDTO struct to the usecase package, which will be used for transferring spot data between layers. It also updates the Event and Ticket models in the domain package to include new fields and methods. The NewEvent function is added to create a new event with the given parameters, and the NewTicket function is added to create a new ticket for an event and spot.